Change type of user type

+++ foxogram-common/src/main/java/su/foxogram/dtos/api/response/UserDTO.java
@@ -29,7 +29,7 @@ public class UserDTO {
 
 	private long flags;
 
-	private long type;
+	private int type;
 
 	private long createdAt;
